Refactor seg into tasks; add CMake options
Introduce build-time OPMODE and SERIAL_MODE CMake options and expose them as compiler defines. Add new serial_to_ethernet sources (seg_proc.c, seg_data.c, seg_util.c) to the app library. Refactor seg code: extract protocol-specific handlers, declare many shared globals and FreeRTOS handles in seg.h, and modularize seg.c to use function-pointer-based handlers for operation mode and serial protocol. Convert seg workflow to RTOS tasks (seg_task, seg_u2e_task, seg_recv_task, seg_timer_task, seg_mqtt_yield_task), add timer callbacks and centralized timer management, and move SPI/UART transfer logic into modular handlers. Also change u2e_size/e2u_size type to int to match usage and update spiHandler externs accordingly.

+++ port/app/platform_handler/src/spiHandler.c
@@ -38,8 +38,8 @@ extern uint8_t g_send_buf[];
 extern uint8_t g_recv_buf[];
 extern uint8_t gSEGCPREQ[];
 extern uint8_t gSEGCPREP[];
-extern uint16_t e2u_size;
-extern uint16_t u2e_size;
+extern int e2u_size;
+extern int u2e_size;
 
 void on_spi_rx(void) {
     //uartRxByte: // 1-byte character variable for UART Interrupt request handler
